improve patients/ structure
