+++ frontend/src/features/patients/components/NewPatientButton.jsx
@@ -2,7 +2,7 @@ import { HiOutlinePlus } from 'react-icons/hi2'
 import Button from '@components/Button'
 import Modal from '@components/Modal'
 import useUser from '@features/users/hooks/useUser'
-import MedicalPatientForm from './MedicalPatientForm/MedicalPatientForm'
+import MedicalPatientForm from '../forms/MedicalPatientForm/MedicalPatientForm'
 
 const AREA_FORMS = {
   medicina: {
