+++ frontend/src/features/patients/forms/MedicalPatientForm/steps/AntecedentesPatologicosStep.jsx
