+++ frontend/src/features/patients/forms/MedicalPatientForm/steps/AparatosSistemasStep.jsx
