+++ frontend/src/features/patients/forms/MedicalPatientForm/steps/ExploracionFisicaStep.jsx
