+++ frontend/src/features/patients/forms/shared/cie10Data.js
