--- frontend/src/features/patients/historia/mockData.js
@@ -1,174 +1,0 @@
-// TODO: Replace with API call via useHistoria hook
-export const HISTORIAS = {
-  '2026-2031': {
-    tipo_sangre: 'O+',
-    vacunas_infancia_completas: true,
-    motivo_consulta:
-      'Consulta de control semestral. Paciente refiere cefalea frecuente de predominio vespertino.',
-    historia_enfermedad_actual:
-      'Paciente de 22 años que acude por cefalea de 2 semanas de evolución, de características tensionales, EVA 6/10, sin irradiación, sin náusea ni fotofobia.',
-    informacion_fisica: {
-      peso: 72.5,
-      altura: 1.75,
-      pa_sistolica: 118,
-      pa_diastolica: 78,
-      fc: 72,
-      fr: 16,
-      circ_cintura: 84.0,
-      circ_cadera: 96.0,
-      sp_o2: 98.5,
-      glucosa_capilar: 94.0,
-      temperatura: 36.7,
-      exploracion_fisica:
-        'Paciente consciente, orientado en tiempo, lugar y persona. Bien hidratado, con adecuado estado general.',
-      habito_exterior:
-        'Buen estado general, cooperador, sin fascies de dolor en reposo.',
-    },
-    antecedentes_patologicos: {
-      cronico_degenerativos: 'Niega',
-      quirurgicos: 'Apendicectomía a los 15 años, sin complicaciones.',
-      hospitalizaciones: 'Niega',
-      traumaticos: 'Fractura de radio derecho a los 10 años, consolidada.',
-      transfusionales: 'Niega',
-      transplantes: null,
-      alergicos: 'Penicilina (urticaria generalizada)',
-      infectocontagiosos: 'Varicela en la infancia.',
-      toxicomanias: null,
-      covid_19: 'COVID-19 en 2021, evolución leve, sin secuelas.',
-      psicologia_psiquiatria: null,
-      gyo: null,
-      enfermedades_congenitas: null,
-      enfermedades_infancia: 'Varicela, sarampión.',
-    },
-    antecedentes_familiares: {
-      padre: 'Hipertensión arterial sistémica, Diabetes mellitus tipo 2',
-      madre: 'Hipotiroidismo',
-      abuelo_paterno: 'Cardiopatía isquémica',
-      abuelo_materno: 'Niega',
-      abuela_paterna: 'Diabetes mellitus tipo 2',
-      abuela_materna: null,
-      otros: null,
-    },
-    aparatos_sistemas: {
-      neurologico: 'Cefalea tensional frecuente.',
-      cardiovascular: 'Sin alteraciones referidas.',
-      respiratorio: 'Sin alteraciones referidas.',
-      hematologico: null,
-      digestivo: 'Reflujo gastroesofágico ocasional.',
-      musculoesqueletico: null,
-      genitourinario: null,
-      endocrinologico: null,
-      metabolico: null,
-      nutricional: null,
-    },
-    inmunizaciones: {
-      influenza: '2025-10-15',
-      tetanos: '2022-03-20',
-      hepatitis_b: '2018-06-10',
-      covid_19: '2021-07-15',
-      otros: null,
-    },
-    planes_estudio: {
-      fecha_generacion: '2026-01-15',
-      cie10_codes: [
-        { code: 'G44.2', description: 'Cefalea tensional' },
-        {
-          code: 'K21.0',
-          description: 'Enfermedad por reflujo gastroesofágico con esofagitis',
-        },
-      ],
-      plan_tratamiento:
-        'Paracetamol 500 mg c/8h por 5 días. Control en 2 semanas si persiste.',
-      tratamiento:
-        'Analgesia, reposo relativo, hidratación adecuada. Evitar pantallas prolongadas.',
-    },
-    servicios: {
-      gas: true,
-      luz: true,
-      agua: true,
-      drenaje: true,
-      cable_tel: false,
-      internet: true,
-    },
-  },
-  '2021-2026': {
-    tipo_sangre: 'O+',
-    vacunas_infancia_completas: true,
-    motivo_consulta: 'Revisión anual. Paciente sin quejas activas.',
-    historia_enfermedad_actual:
-      'Paciente acude a revisión de rutina, sin síntomas activos relevantes.',
-    informacion_fisica: {
-      peso: 70.0,
-      altura: 1.75,
-      pa_sistolica: 115,
-      pa_diastolica: 75,
-      fc: 68,
-      fr: 15,
-      circ_cintura: 82.0,
-      circ_cadera: 94.0,
-      sp_o2: 99.0,
-      glucosa_capilar: 90.0,
-      temperatura: 36.5,
-      exploracion_fisica: 'Sin hallazgos patológicos.',
-      habito_exterior: 'Buen estado general.',
-    },
-    antecedentes_patologicos: {
-      cronico_degenerativos: 'Niega',
-      quirurgicos: 'Apendicectomía a los 15 años.',
-      hospitalizaciones: 'Niega',
-      traumaticos: 'Fractura de radio derecho a los 10 años.',
-      transfusionales: 'Niega',
-      transplantes: null,
-      alergicos: 'Penicilina',
-      infectocontagiosos: 'Varicela en la infancia.',
-      toxicomanias: null,
-      covid_19: null,
-      psicologia_psiquiatria: null,
-      gyo: null,
-      enfermedades_congenitas: null,
-      enfermedades_infancia: 'Varicela, sarampión.',
-    },
-    antecedentes_familiares: {
-      padre: 'Hipertensión arterial sistémica',
-      madre: 'Hipotiroidismo',
-      abuelo_paterno: null,
-      abuelo_materno: 'Niega',
-      abuela_paterna: 'Diabetes mellitus tipo 2',
-      abuela_materna: null,
-      otros: null,
-    },
-    aparatos_sistemas: {
-      neurologico: null,
-      cardiovascular: 'Sin alteraciones.',
-      respiratorio: 'Sin alteraciones.',
-      hematologico: null,
-      digestivo: null,
-      musculoesqueletico: null,
-      genitourinario: null,
-      endocrinologico: null,
-      metabolico: null,
-      nutricional: null,
-    },
-    inmunizaciones: {
-      influenza: '2021-11-02',
-      tetanos: '2019-05-14',
-      hepatitis_b: '2018-06-10',
-      covid_19: null,
-      otros: null,
-    },
-    planes_estudio: {
-      fecha_generacion: '2022-03-10',
-      cie10_codes: [{ code: 'Z00.0', description: 'Examen médico general' }],
-      plan_tratamiento: 'Sin indicaciones especiales.',
-      tratamiento: null,
-    },
-    servicios: {
-      gas: true,
-      luz: true,
-      agua: true,
-      drenaje: true,
-      cable_tel: false,
-      internet: false,
-    },
-  },
-}
