+++ frontend/src/features/patients/pages/PatientDetail.jsx
@@ -4,15 +4,15 @@ import { HiOutlineTrash } from 'react-icons/hi2'
 import Modal from '@components/Modal'
 import DangerConfirm from '@components/DangerConfirm'
 import Tab from '@components/Tab'
-import { usePatient } from './hooks/usePatient'
-import { useDeletePatient } from './hooks/useDeletePatient'
-import PatientActionBar from './components/PatientActionBar'
-import PatientHeader from './components/PatientHeader'
-import NotesPanel from './components/NotesPanel'
-import PersonalDataPanel from './components/PersonalDataPanel'
-import PatientSkeleton from './components/PatientSkeleton'
-import PatientHistoria from './PatientHistoria'
-import MedicalPatientForm from './components/MedicalPatientForm/MedicalPatientForm'
+import { usePatient } from '../hooks/usePatient'
+import { useDeletePatient } from '../hooks/useDeletePatient'
+import PatientActionBar from '../components/PatientActionBar'
+import PatientHeader from '../components/PatientHeader'
+import NotesPanel from '../components/NotesPanel'
+import PersonalDataPanel from '../components/PersonalDataPanel'
+import PatientSkeleton from '../components/PatientSkeleton'
+import PatientHistoria from '../components/PatientHistoria'
+import MedicalPatientForm from '../forms/MedicalPatientForm/MedicalPatientForm'
 
 export default function PatientDetail() {
   const { patient, isPending } = usePatient()
