+++ frontend/src/features/patients/pages/PatientRow.jsx
@@ -4,7 +4,7 @@ import Modal from '@components/Modal'
 import RowActionsMenu from '@components/RowActionsMenu'
 import Table from '@components/Table'
 import { HiOutlineTrash, HiArrowRight } from 'react-icons/hi2'
-import { useDeletePatient } from './hooks/useDeletePatient'
+import { useDeletePatient } from '../hooks/useDeletePatient'
 import { useNavigate } from 'react-router'
 import DateTime from '@components/DateTime'
 import BirthDate from '@components/BirthDate'
