+++ frontend/src/features/patients/sections/FieldsSection.jsx
@@ -1,6 +1,6 @@
 import Grid from '@components/Grid'
-import DataField from '../../components/DataField'
-import Empty from '../components/Empty'
+import DataField from '@components/DataField'
+import Empty from './Empty'
 
 export default function FieldsSection({ fields, cols = 2, mobileCols }) {
   if (!fields) return <Empty />
