+++ frontend/src/features/patients/sections/NoPatologicosSection.jsx
@@ -1,10 +1,7 @@
-import {
-  HiOutlineBeaker,
-  HiOutlineCheckCircle,
-  HiOutlineXCircle,
-} from 'react-icons/hi2'
+import { HiOutlineBeaker } from 'react-icons/hi2'
 import Heading from '@components/Heading'
-import DataField from '../../components/DataField'
+import DataField from '@components/DataField'
+import StatusChip from '@components/StatusChip'
 import InmunizacionesSection from './InmunizacionesSection'
 
 const SERVICIOS_ITEMS = [
@@ -33,27 +30,6 @@ const ANTECEDENTES_ITEMS = [
     labelFalse: 'Sin zoonosis',
   },
 ]
-
-function StatusChip({ active, children, size = 'sm' }) {
-  const sizeClasses = size === 'sm' ? 'text-6 px-3 py-1' : 'text-5 px-4 py-2'
-  const iconSize = size === 'sm' ? 12 : 14
-  return (
-    <span
-      className={`inline-flex items-center gap-1.5 rounded-full border font-medium ${sizeClasses} ${
-        active
-          ? 'border-green-100 bg-green-50 text-green-700'
-          : 'border-zinc-200 bg-zinc-100 text-zinc-400'
-      }`}
-    >
-      {active ? (
-        <HiOutlineCheckCircle size={iconSize} />
-      ) : (
-        <HiOutlineXCircle size={iconSize} />
-      )}
-      {children}
-    </span>
-  )
-}
 
 export default function NoPatologicosSection({ historia }) {
   const {
