+++ frontend/src/pages/PatientDetail.jsx
@@ -1,4 +1,4 @@
-import PatientDetail from '@features/patients/PatientDetail'
+import PatientDetail from '@features/patients/pages/PatientDetail'
 
 export default function PatientDetailPage() {
   return <PatientDetail />
